+++ packages/falcon-client/src/middlewares/routes/ssrMiddleware.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { StaticRouter } from 'react-router-dom';
-import { ApolloProvider, getDataFromTree } from 'react-apollo';
+import { ApolloProvider } from '@apollo/react-common';
+import { getDataFromTree } from '@apollo/react-ssr';
 import Helmet from 'react-helmet';
 import { ChunkExtractorManager } from '@loadable/server';
 import { I18nProvider } from '@deity/falcon-i18n';