+++ packages/falcon-shop-data/src/Cart/AddToCartMutation.tsx
@@ -1,6 +1,5 @@
 import gql from 'graphql-tag';
-import { Mutation } from 'react-apollo';
-import { OperationInput } from '@deity/falcon-data';
+import { Mutation, OperationInput } from '@deity/falcon-data';
 import { AddToCartInput, CartItemPayload } from '@deity/falcon-shop-extension';
 
 export const ADD_TO_CART = gql`